feat(cal-ui-p1): repoint app calendar to package (shims + extracted time-window helpers)
